backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
asyncer==0.0.8
attrs==25.3.0
autogen-agentchat==0.6.1
autogen-core==0.6.1
autogen-ext==0.6.1
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
decorator==5.2.1
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docopt==0.6.2
et-xmlfile==2.0.0
executing==2.2.0
fastapi==0.110.0
fastjsonschema==2.21.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==8.7.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.3.1
openai==1.93.0
openpyxl==3.1.2
opentelemetry-api==1.34.1
packaging==25.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.3.8
prompt-toolkit==3.0.51
protobuf==5.29.5
pure-eval==0.2.3
pydantic==2.11.7
pydantic-core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.2
pywin32==310
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.25.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.36.3
termcolor==3.1.0
tiktoken==0.9.0
tinycss2==1.4.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing-extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.29.0
wcwidth==0.2.13
webencodings==0.5.1
XlsxWriter==3.2.0
yarg==0.1.9
zipp==3.23.0